+++ frontend/src/pages/TransactionsPage.js
@@ -10,13 +10,17 @@ import {
 
 import { Card, ProgressBar, Stack, Button } from 'react-bootstrap'
 
-import { Formik, Field, Form, ErrorMessage } from 'formik'
+import { Formik, Field, Form, ErrorMessage, useField } from 'formik'
 import * as Yup from 'yup'
 
 import { currencyFormatter } from '../utils/tools'
 
 import Layout from '../components/Layout/Layout'
 import Transaction from '../components/Transaction/Transaction'
+
+// picking dates
+import DatePicker from 'react-datepicker'
+import 'react-datepicker/dist/react-datepicker.css'
 
 function TransactionsPage() {
   const params = useParams()
@@ -30,8 +34,13 @@ function TransactionsPage() {
       .min(3, 'text must be at least 3 characters long')
       .max(75, 'text must be less than 75 characters long')
       .required('Required'),
+    category: Yup.string()
+      .min(3, 'categories must be at least 3 characters long')
+      .max(75, 'categories must be less than 75 characters long')
+      .required('Required'),
 
     amount: Yup.number().default(0).required('Required'),
+    date: Yup.string().required('Required'),
   })
 
   const submitForm = (values) => {
@@ -80,23 +89,49 @@ function TransactionsPage() {
     return 'danger'
   }
 
+  // categories:
+  const allCategories = [
+    'all',
+    ...new Set(transactions.map((transaction) => transaction.category)),
+  ]
+
   useEffect(() => {
     dispatch(getTransactions(params.id))
   }, [])
+  const [categories, setCategories] = useState(allCategories)
+  const [newTransactions, setNewTransactions] = useState(transactions)
 
-  const handleDeletion = (id) => {
-    console.log(id)
-    dispatch(deleteTransaction(id))
-      .unwrap()
-      .then(() => {
-        console.log('done!')
-        setTimeout(() => window.location.reload(), 1500)
-      })
-      .catch(() => {
-        setError(true)
-        console.log('error')
-      })
+  // picking date stuff
+  const [startDate, setStartDate] = useState(new Date())
+  const MyDatePicker = ({ name = '' }) => {
+    const [field, meta, helpers] = useField(name)
+
+    const { value } = meta
+    const { setValue } = helpers
+
+    return (
+      <DatePicker
+        {...field}
+        selected={value}
+        onChange={(date) => setValue(date)}
+      />
+    )
   }
+
+  const filterItems = (category) => {
+    if (category === 'all') {
+      setNewTransactions(transactions) /* og array */
+      return
+    }
+    /* iterating over og list */
+    const newItems = transactions.filter(
+      (transactions) => transactions.category === category
+    )
+    setNewTransactions(newItems)
+  }
+  console.log(allCategories)
+  console.log('categories', categories)
+  console.log('newTransactions', newTransactions)
 
   return (
     <Layout>
@@ -107,25 +142,32 @@ function TransactionsPage() {
             user: params.id,
             text: '',
             amount: '',
+            date: new Date(),
           }}
           validationSchema={ValidationSchema}
           onSubmit={(values, { setSubmitting }) => {
             setTimeout(() => {
+              console.log(values)
               submitForm(values)
               setSubmitting(false)
-            }, 400)
+            }, 500)
           }}
         >
-          {({ isSubmitting }) => (
+          {({ values, isSubmitting }) => (
             <Form>
               <div className='d-flex flex-column w-50'>
                 <>
-                  <Field type='text' name='text' placeholder='your text' />
+                  <Field type='text' name='text' placeholder='transaction' />
                   <ErrorMessage name='text' component='div' />
+                </>
+                <>
+                  <Field type='text' name='category' placeholder='category' />
+                  <ErrorMessage name='category' component='div' />
                 </>
 
                 <Field type='number' name='amount' placeholder='amount' />
                 <ErrorMessage name='amount' component='div' />
+                <MyDatePicker name='date' />
               </div>
 
               <div className='flex flex-col items-start'>
@@ -152,21 +194,29 @@ function TransactionsPage() {
           transactions.map((transaction) => (
             <>
               <Transaction key={transaction._id} transaction={transaction} />
-              <button onClick={() => handleDeletion(transaction._id)}>
-                delete {transaction._id}
-              </button>
             </>
           ))
         )}
       </div>
-      <div className='bg-success my-2 p-3'>
-        <h2>income</h2>
-        <span>{currencyFormatter.format(income)}</span>
+      <div className='btn-container'>
+        {categories.map((category, index) => (
+          <button
+            key={index}
+            className='filter-btn'
+            onClick={() => filterItems(category)}
+          >
+            {category}
+          </button>
+        ))}
       </div>
-      <div className='bg-danger my-2 p-3'>
+      <div className='bg-success my-2 p-3 d-flex align-items-center justify-content-between'>
+        <h2>income</h2>
+        <span className='fs-3'>{currencyFormatter.format(income)}</span>
+      </div>
+      <div className='bg-danger my-2 p-3 d-flex align-items-center justify-content-between'>
         <h2>expenses</h2>
 
-        <span>{currencyFormatter.format(expense)}</span>
+        <span className='fs-3'>{currencyFormatter.format(expense)}</span>
       </div>
       <div>
         <h2>total</h2>